Format StmtAssign, badly
This is a near-verbatim formatting of assignment statements so we can see the effect of formatting the right hand side. It lacks the wrapping rules for the left hand side and will put the entire left hand side on one line.

On the snapshots this has mostly the effect of normalizing the whitespace around equals sign, but more importantly it will surface all changes to expressions.