Suggest combining async with statements
Previously the rule for SIM117 explicitly ignored async with statements
as it would incorrectly suggestion to merge async with and regular with
statements as reported in issue #1902.

This partially reverts the fix for that (commit
396be5edeac0c5724de87e93c5a885dacf201f05) by enabling the rules for
async with statements again, but with a check ensuring that the
statements are both of the same kind, i.e. both async with or both
(just) with statements.