Unify benchmarking and profiling docs
This moves all docs about benchmarking and profiling into CONTRIBUTING.md by moving the readme of `ruff_benchmark` and adding more information on profiling.

We need to somehow consolidate that documentation, but i'm not convinced that this is the best way (i tried subpages in mkdocs, but that didn't seem good either), so i'm happy to take suggestions.