Split PackageCache into Cache
Cache will be the type managing the cache (file), while PackageCache now
only holds the on-disk data.

This changes the way how the cache is created and updated. The new Cache
type keeps track of the source file (and cache) that changed, meaning
that if the source file didn't change we don't have to do much (other
than printing the diagnostics to the user). This means we can remove the
lock from PackageCache::files, making reading the cache files cheaper.
Of course we still need a lock around Cache::new_files to keep track of
the files that have changed.

But working with the assumption that changed source file are smaller in
number than the unchanged source files this leads to less contention on
the hashmap lock (that now has been reduced in scope).

This, along side the previous improvements, reduce a fully cached run to
be ~10% slower (up from 50%+ slower) than the old cache implementation,
that used a cache file per source file resulting in larger cache sizes.