Keep track of when files are last seen in the cache
And remove cached files that we haven't seen for a certain period of
time, currently 30 days.

For the last seen timestamp we actually use an u64, it's smaller on disk
than SystemTime (which size is OS dependant) and fits in an AtomicU64
which we can use to update it without locks.